minor cleaning of the trainig_loop folder (moving imbalance.py to optimisation folder) and grouped tests by folder to make navigation easier.
