+++ modules/optimisation/__init__.py
@@ -1,11 +1,13 @@
 """Optimisation utilities for optimisers, learning-rate schedulers and loss functions."""
 
-from .scheduler_factory import normalise_scheduler_config, create_scheduler
+from .imbalance import make_weighted_sampler
 from .loss import get_loss_function, FocalLoss
+from .scheduler_factory import normalise_scheduler_config, create_scheduler
 
 __all__ = [
     "normalise_scheduler_config",
     "create_scheduler",
     "get_loss_function",
     "FocalLoss",
+    "make_weighted_sampler",
 ]
