+++ modules/training_loop/__init__.py
@@ -9,6 +9,8 @@ __all__ = ["training", "_build_train_config", "train_model_loop"]
 TRAINING LOOP MODULE // Files related to the main training loop and its components
 - config: 
         Function to build the training configuration dictionary, which includes all necessary parameters and objects for training
+- evaluate:
+        Function that evaluates the model on the test set and returns the test loss and metrics, including checks against client performance requirements
 - one_epoch: 
         Function that trains the model for one epoch and returns the training loss and accuracy
 - validation: 
@@ -18,28 +20,13 @@ TRAINING LOOP MODULE // Files related to the main training loop and its componen
 - run_saving: 
         Functions for initializing the training history, appending metrics to the history, and saving the best model and run summary to disk at the end of training
 - utility: 
-        Utility functions for 
+        Utility functions for... 
                 - safely getting the class name of an object 
                 - converting various types of values (including tensors) to a format that can be easily saved in JSON or similar formats
                 - unpacking batches from the dataloader and preparing them for model input
                 - getting the learning rates from the optimizer's parameter groups
                 - comparing the current metric value to the best metric value based on the specified mode (min or max)
-
-
-Functionality notes (): 
- - Support for different loss functions 
-    -> currently supports any loss function that can be called as `criterion(preds, labels)` and returns a scalar loss value. 
- - More detailed model saving 
-    -> model saving is now more detailed
- - More detailed logging (e.g., using TensorBoard or a logging library instead of print statements)
-    -> currently uses print statements for logging, but could be extended to use a more robust logging framework or TensorBoard for better visualization of training progress and metrics
- - Support for resuming training from a checkpoint
-    -> currently does not support resuming training from a checkpoint, 
-        -> could be extended to allow loading a saved model state dict and training history to continue training from where it left off
- - More flexible learning rate scheduling (e.g., support for different schedulers, or custom scheduling logic)
-    -> currently supports a default StepLR scheduler, or use of a custom scheduler if provided. 
-        -> could be extended to support a wider range of built-in schedulers, or allow for custom scheduling logic to be implemented by the user
- - ....etc.
+                - normalising a class dictionary to ensure keys are integers and values are strings
 """
 
 
@@ -89,6 +76,7 @@ def training(
     #
     extra_config=None,
     requirements={},
+    #
     log_leaderboard=True,
     leaderboard_dir="leaderboard",
     verbose=True,
@@ -102,7 +90,7 @@ def training(
         model_type:     Label used for saving and logging.
         need_length:    Whether the model expects sequence lengths as input.
 
-        optimiser:  Optimizer instance. Defaults to Adam with lr=1e-3.
+        optimiser:      Optimizer instance. Defaults to Adam with lr=1e-3.
         optimiser_args: Dictionary of additional arguments for the optimizer.
 
         scheduler:                  Learning rate scheduler. Defaults to StepLR.
@@ -115,8 +103,8 @@ def training(
         train_dl:  DataLoader for training data.
         valid_dl:  DataLoader for validation data.
         test_dl:   DataLoader for test data.
-        use_weighted_sampler: If True, use WeightedRandomSampler to handle class imbalance. Defaults to False.
-        train_labels: List or tensor of training labels. Required if use_weighted_sampler is True.
+        use_weighted_sampler:   Bool -> If True, use WeightedRandomSampler to handle class imbalance. Defaults to False.
+        train_labels:           List or tensor of training labels. Required if use_weighted_sampler is True.
 
         epochs:              Number of training epochs.
         patience:            Early stopping patience in epochs.
@@ -134,7 +122,7 @@ def training(
         parameters:  The training parameters in a dictionary format.
         device:      Device string, e.g. 'cpu' or 'cuda'.
 
-        compute_train_metrics:  Whether to compute metrics on the training set.
+        compute_train_metrics:  Whether to compute metrics on the training set. Defaults to True.
         save:                   Whether to save model artifacts after training.
         parent_dir:             Directory to save model artifacts.
         run_name:               Optional name for the run.
@@ -142,13 +130,14 @@ def training(
         extra_config:  Optional dict of additional config keys to merge.
         requirements:  Optional client performance requirements dict, defaults to {}. 
             Pass None to disable check. Keys:
-            - confidence_threshold: {"high": float, "medium": float} (values >1 treated as %)
-            - high_threshold: min fraction of predictions in high-confidence tier (default 0.70)
-            - fatal_accuracy: min recall on true fatal-class samples (default 0.95)
-            - f1_target: {class_index: min_f1} — use 0.0 to mark a class as having no target
-        log_leaderboard: Whether to append this run to the leaderboard CSV. Defaults to True.
-        leaderboard_dir: Directory for leaderboard.csv and owner.conf. Defaults to 'leaderboard'.
-        verbose: Enable printing the training loop message. Defaults to True.
+                - confidence_threshold: {"high": float, "medium": float} (values >1 treated as %)
+                - high_threshold: min fraction of predictions in high-confidence tier (default 0.70)
+                - fatal_accuracy: min recall on true fatal-class samples (default 0.95)
+                - f1_target: {class_index: min_f1} — use 0.0 to mark a class as having no target
+        
+        log_leaderboard:    Whether to append this run to the leaderboard CSV. Defaults to True. // If True, requires save to be True for logging.
+        leaderboard_dir:    Directory for leaderboard.csv and owner.conf. Defaults to 'leaderboard'.
+        verbose:            Enable printing the training loop message. Defaults to True.
 
     Returns:
         Run summary dictionary with history, best epoch, and best metric value.
